+++ local/src/main/java/com/kdjj/local/dataSource/LocalDataSource.kt
@@ -1,4 +1,4 @@
-package com.kdjj.local
+package com.kdjj.local.dataSource
 
 import com.kdjj.domain.model.Recipe
