+++ local/src/main/java/com/kdjj/local/dataSource/LocalDataSourceImpl.kt
@@ -1,6 +1,7 @@
-package com.kdjj.local
+package com.kdjj.local.dataSource
 
 import com.kdjj.domain.model.Recipe
+import com.kdjj.local.dataSource.LocalDataSource
 
 class LocalDataSourceImpl: LocalDataSource {
